move dependency management to own file
